panelType replaces modal
