+++ packages/dataviews/CHANGELOG.md
@@ -11,7 +11,7 @@
 - Introduce a new `array` DataForm Edit control that supports multi-selection. [#71136](https://github.com/WordPress/gutenberg/pull/71136)
 - Add `enableMoving` option to the `table` layout to allow or disallow column moving left and right. [#71120](https://github.com/WordPress/gutenberg/pull/71120)
 - Add infinite scroll support across all layout types (grid, list, table). Enable infinite scroll by providing an `infiniteScrollHandler` function in the `paginationInfo` prop and toggling the feature in the view configuration. ([#70955](https://github.com/WordPress/gutenberg/pull/70955))
-- Add support for modal DataForm layout and control over the content of combined fields. [#71212](https://github.com/WordPress/gutenberg/pull/71212)
+- Add support for modal in DataForm panel layouts, and control over the content of combined fields. [#71212](https://github.com/WordPress/gutenberg/pull/71212)
 
 ### Enhancements
