+++ packages/dataviews/src/dataforms-layouts/data-form-layout.tsx
@@ -48,13 +48,7 @@ export function DataFormLayout< Item >( {
 	);
 
 	return (
-		<VStack
-			spacing={
-				form.layout?.type === 'panel' || form.layout?.type === 'modal'
-					? 2
-					: 4
-			}
-		>
+		<VStack spacing={ form.layout?.type === 'panel' ? 2 : 4 }>
 			{ normalizedFormFields.map( ( formField ) => {
 				const FieldLayout = getFormFieldLayout( formField.layout.type )
 					?.component;
