+++ packages/dataviews/src/dataforms-layouts/index.tsx
@@ -4,7 +4,6 @@
 import FormRegularField from './regular';
 import FormPanelField from './panel';
 import FormCardField from './card';
-import FormModalField from './modal';
 
 const FORM_FIELD_LAYOUTS = [
 	{
@@ -18,10 +17,6 @@ const FORM_FIELD_LAYOUTS = [
 	{
 		type: 'card',
 		component: FormCardField,
-	},
-	{
-		type: 'modal',
-		component: FormModalField,
 	},
 ];
