--- packages/dataviews/src/dataforms-layouts/modal/index.tsx
@@ -1,260 +1,0 @@
-/**
- * External dependencies
- */
-import clsx from 'clsx';
-
-/**
- * WordPress dependencies
- */
-import {
-	__experimentalVStack as VStack,
-	__experimentalHStack as HStack,
-	__experimentalSpacer as Spacer,
-	Button,
-	Modal,
-} from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
-import { useState, useMemo, useContext } from '@wordpress/element';
-
-/**
- * Internal dependencies
- */
-import type { Form, FieldLayoutProps, SimpleFormField } from '../../types';
-import DataFormContext from '../../components/dataform-context';
-import { DataFormLayout } from '../data-form-layout';
-import { isCombinedField } from '../is-combined-field';
-
-function DataFormModalFooter( {
-	onCancel,
-	onApply,
-}: {
-	onClose: () => void;
-	onCancel: () => void;
-	onApply: () => void;
-} ) {
-	return (
-		<HStack className="dataforms-layouts-modal__modal-footer" spacing={ 3 }>
-			<Spacer />
-			<Button
-				variant="tertiary"
-				onClick={ onCancel }
-				__next40pxDefaultSize
-			>
-				{ __( 'Cancel' ) }
-			</Button>
-			<Button variant="primary" onClick={ onApply } __next40pxDefaultSize>
-				{ __( 'Apply' ) }
-			</Button>
-		</HStack>
-	);
-}
-
-function DataFormModal< Item >( {
-	isOpen,
-	onClose,
-	data,
-	form,
-	onChange,
-	fieldLabel,
-	onCancel,
-	onApply,
-}: {
-	isOpen: boolean;
-	onClose: () => void;
-	data: Item;
-	form: Form;
-	onChange: ( data: Partial< Item > ) => void;
-	onCancel: () => void;
-	onApply: () => void;
-	fieldLabel: string;
-} ) {
-	if ( ! isOpen ) {
-		return null;
-	}
-
-	return (
-		<Modal
-			className="dataforms-layouts-modal__modal"
-			onRequestClose={ onClose }
-			isFullScreen={ false }
-			title={ fieldLabel }
-			shouldCloseOnClickOutside={ false }
-		>
-			<DataFormLayout data={ data } form={ form } onChange={ onChange }>
-				{ ( FieldLayout, nestedField ) => (
-					<FieldLayout
-						key={ nestedField.id }
-						data={ data }
-						field={ nestedField }
-						onChange={ onChange }
-						hideLabelFromVision={
-							( form?.fields ?? [] ).length < 2
-						}
-					/>
-				) }
-			</DataFormLayout>
-			<DataFormModalFooter
-				onClose={ onClose }
-				onCancel={ onCancel }
-				onApply={ onApply }
-			/>
-		</Modal>
-	);
-}
-
-export default function FormModalField< Item >( {
-	data,
-	field,
-	onChange,
-}: FieldLayoutProps< Item > ) {
-	const { fields } = useContext( DataFormContext );
-	const [ isOpen, setIsOpen ] = useState( false );
-	const [ localData, setLocalData ] = useState< Item >( data );
-
-	const fieldDefinition = fields.find( ( fieldDef ) => {
-		// Default to the first child if it is a combined field.
-		if ( isCombinedField( field ) ) {
-			const children = field.children.filter(
-				( child ): child is string | SimpleFormField =>
-					typeof child === 'string' || ! isCombinedField( child )
-			);
-			const firstChildFieldId =
-				typeof children[ 0 ] === 'string'
-					? children[ 0 ]
-					: children[ 0 ].id;
-			// Return first child regardless of field id.
-			return !! firstChildFieldId;
-		}
-		return fieldDef.id === field.id;
-	} );
-
-	const form = useMemo( () => {
-		if ( isCombinedField( field ) ) {
-			return {
-				type: 'regular' as const,
-				fields: field.children.map( ( child ) => {
-					if ( typeof child === 'string' ) {
-						return {
-							id: child,
-						};
-					}
-					return child;
-				} ),
-			};
-		}
-		// If not explicit children return the field id itself.
-		return {
-			type: 'regular' as const,
-			fields: [ { id: field.id } ],
-		};
-	}, [ field ] );
-
-	if ( ! fieldDefinition ) {
-		return null;
-	}
-
-	const labelPosition = field.labelPosition ?? 'side';
-	const labelClassName = clsx(
-		'dataforms-layouts-modal__field-label',
-		`dataforms-layouts-modal__field-label--label-position-${ labelPosition }`
-	);
-
-	const fieldLabel = isCombinedField( field )
-		? field.label
-		: fieldDefinition?.label;
-
-	const handleOpenModal = () => {
-		setIsOpen( true );
-		// Reset local data to current data when opening modal
-		setLocalData( data );
-	};
-
-	const handleCloseModal = () => {
-		setIsOpen( false );
-		// Close button should also cancel changes
-		setLocalData( data );
-	};
-
-	const handleCancelModal = () => {
-		setIsOpen( false );
-		// Discard changes by resetting local data
-		setLocalData( data );
-	};
-
-	const handleApplyModal = () => {
-		setIsOpen( false );
-		// Apply changes when Apply button is clicked
-		onChange( localData );
-	};
-
-	const handleLocalChange = ( value: Partial< Item > ) => {
-		// Update local state without calling onChange
-		setLocalData( ( prev ) => ( { ...prev, ...value } ) );
-	};
-
-	const renderedControl = (
-		<>
-			<Button
-				className="dataforms-layouts-modal__field-button"
-				size="compact"
-				variant="link"
-				onClick={ handleOpenModal }
-				disabled={ fieldDefinition.readOnly === true }
-				accessibleWhenDisabled
-			>
-				{ isCombinedField( field ) && field?.content ? (
-					field.content
-				) : (
-					<fieldDefinition.render
-						item={ data }
-						field={ fieldDefinition }
-					/>
-				) }
-			</Button>
-			<DataFormModal
-				isOpen={ isOpen }
-				onClose={ handleCloseModal }
-				onCancel={ handleCancelModal }
-				onApply={ handleApplyModal }
-				data={ localData }
-				form={ form as Form }
-				fieldLabel={ fieldLabel ?? '' }
-				onChange={ handleLocalChange }
-			/>
-		</>
-	);
-
-	if ( labelPosition === 'top' ) {
-		return (
-			<VStack className="dataforms-layouts-modal__field" spacing={ 0 }>
-				<div
-					className={ labelClassName }
-					style={ { paddingBottom: 0 } }
-				>
-					{ fieldLabel }
-				</div>
-				<div className="dataforms-layouts-modal__field-control">
-					{ renderedControl }
-				</div>
-			</VStack>
-		);
-	}
-
-	if ( labelPosition === 'none' ) {
-		return (
-			<div className="dataforms-layouts-modal__field">
-				{ renderedControl }
-			</div>
-		);
-	}
-
-	// Defaults to label position side.
-	return (
-		<HStack className="dataforms-layouts-modal__field">
-			<div className={ labelClassName }>{ fieldLabel }</div>
-			<div className="dataforms-layouts-modal__field-control">
-				{ renderedControl }
-			</div>
-		</HStack>
-	);
-}
