--- packages/dataviews/src/dataforms-layouts/modal/style.scss
@@ -1,57 +1,0 @@
-.dataforms-layouts-modal__field {
-	width: 100%;
-	min-height: $grid-unit-40;
-	justify-content: flex-start !important;
-	align-items: flex-start !important;
-}
-
-.dataforms-layouts-modal__field-label {
-	width: 38%;
-	flex-shrink: 0;
-	min-height: $grid-unit-40;
-	display: flex;
-	align-items: center;
-	line-height: $grid-unit-05 * 5;
-	hyphens: auto;
-
-	&--label-position-side {
-		align-self: center;
-	}
-}
-
-.dataforms-layouts-modal__field-control {
-	flex-grow: 1;
-	min-height: $grid-unit-40;
-	display: flex;
-	align-items: center;
-
-	.components-button {
-		max-width: 100%;
-		text-align: left;
-		white-space: normal;
-		text-wrap: balance; // Fallback for Safari.
-		text-wrap: pretty;
-		min-height: $button-size-compact;
-	}
-
-	&.components-button.is-link[aria-disabled="true"] {
-		text-decoration: none;
-	}
-}
-
-.dataforms-layouts-modal__field-button {
-	max-width: 100%;
-	text-align: left;
-	white-space: normal;
-	text-wrap: balance; // Fallback for Safari.
-	text-wrap: pretty;
-	min-height: $button-size-compact;
-}
-
-.dataforms-layouts-modal__modal-header {
-	margin-bottom: $grid-unit-20;
-}
-
-.dataforms-layouts-modal__modal-footer {
-	margin-top: $grid-unit-20;
-}
