+++ packages/dataviews/src/style.scss
@@ -14,4 +14,3 @@
 @import "./dataform-controls/style.scss";
 @import "./dataforms-layouts/panel/style.scss";
 @import "./dataforms-layouts/regular/style.scss";
-@import "./dataforms-layouts/modal/style.scss";
