+++ packages/dataviews/src/types.ts
@@ -689,19 +689,14 @@ export type NormalizedRegularLayout = {
 export type PanelLayout = {
 	type: 'panel';
 	labelPosition?: LabelPosition;
+	panelType?: 'dropdown' | 'modal';
 };
 export type NormalizedPanelLayout = {
 	type: 'panel';
 	labelPosition: LabelPosition;
+	panelType: 'dropdown' | 'modal';
 };
-export type ModalLayout = {
-	type: 'modal';
-	labelPosition?: LabelPosition;
-};
-export type NormalizedModalLayout = {
-	type: 'modal';
-	labelPosition: LabelPosition;
-};
+
 export type CardLayout =
 	| {
 			type: 'card';
@@ -729,12 +724,11 @@ export type NormalizedCardLayout =
 			isOpened: boolean;
 	  };
 
-export type Layout = RegularLayout | PanelLayout | CardLayout | ModalLayout;
+export type Layout = RegularLayout | PanelLayout | CardLayout;
 export type NormalizedLayout =
 	| NormalizedRegularLayout
 	| NormalizedPanelLayout
-	| NormalizedCardLayout
-	| NormalizedModalLayout;
+	| NormalizedCardLayout;
 
 export type SimpleFormField = {
 	id: string;
